20170411_1455 - Fixed a bug where an API call could enter unintended loop
Replaced the followup message to Watson Conversation with a blank
string instead of the original message content. If the calling node was
at the root level with no follow up node, it was possible for the
message to end up back at the original calling node and continuing to
loop.